Fix missing attribute and move to another package
